requirements.txt
pydantic
fastapi[standard]==0.119.0
uvicorn[standard]==0.37.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
